backend/requirements.txt
# Requirements para Backend - Todos los Agentes
# ==============================================
# Este archivo incluye todas las dependencias necesarias para
# el backend completo, incluyendo todos los sub-agentes

# ============================================================================
# CORE - LangChain & LangGraph
# ============================================================================
langchain>=0.1.0
langgraph>=0.0.20
langgraph-checkpoint-postgres>=1.0.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# ============================================================================
# LLM PROVIDERS
# ============================================================================
langchain-anthropic>=0.1.0  # Claude Haiku 3
# langchain-openai>=0.0.5   # Descomentar si usas OpenAI

# ============================================================================
# EMBEDDINGS (Local)
# ============================================================================
sentence-transformers>=2.2.0
torch==2.9.0  # CPU version - compatible con sentence-transformers
# Nota: Instalar con: pip install torch==2.5.0 --index-url https://download.pytorch.org/whl/cpu

# ============================================================================
# VECTOR STORE & DATABASE
# ============================================================================
pgvector>=0.2.0
asyncpg>=0.29.0  # PostgreSQL async driver (backup)
psycopg[binary]>=3.1.0  # PostgreSQL driver para langgraph checkpoint
psycopg-pool>=3.1.0  # Connection pool para psycopg
psycopg2-binary>=2.9.0  # PostgreSQL sync driver (backup)

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6  # Para file uploads

# ============================================================================
# CONFIGURATION & ENV
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# HTTP & REQUESTS
# ============================================================================
httpx>=0.25.0  # Cliente HTTP async
aiohttp>=3.9.0  # Cliente HTTP async alternativo

# ============================================================================
# METRICS & MONITORING
# ============================================================================
prometheus-client>=0.19.0

# ============================================================================
# LOGGING
# ============================================================================
structlog>=24.0.0  # Structured logging
python-json-logger>=2.0.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dateutil>=2.8.0
pytz>=2024.1

# ============================================================================
# TESTING
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-httpx>=0.21.0  # Para mockear requests HTTP

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
black>=23.0.0  # Code formatter
flake8>=6.0.0  # Linter
mypy>=1.7.0  # Type checker
isort>=5.12.0  # Import sorter
pre-commit>=3.5.0  # Git hooks

# ============================================================================
# JUPYTER (Opcional - para desarrollo)
# ============================================================================
jupyter>=1.0.0
ipython>=8.0.0
notebook>=7.0.0

# ============================================================================
# WHATSAPP (Si usas WhatsApp Web.js desde Python)
# ============================================================================
# Nota: WhatsApp Web.js es Node.js, pero si necesitas integrarlo:
# websockets>=12.0  # Para comunicación con Node.js

# ============================================================================
# EXTRAS (Descomentar según necesidad)
# ============================================================================
# redis>=5.0.0  # Para caching
# celery>=5.3.0  # Para tareas asíncronas
# sqlalchemy>=2.0.0  # ORM (si prefieres ORM sobre SQL directo)
# alembic>=1.12.0  # Migraciones de BD
